Refactor MixFP4 compressor helpers
