+++ src/compressed_tensors/compressors/mixfp4/__init__.py
@@ -4,3 +4,4 @@
 # flake8: noqa
 
 from .base import *
+from .helpers import *
